agentic-research-system/requirements.txt
# Core & Scheduling
apscheduler==3.10.4
python-dotenv==1.0.0

# Data Extraction
requests==2.31.0
feedparser==6.0.10
sec-api==1.0.32
aiohttp>=3.11.11

# Web Scraping
playwright==1.40.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.2
lxml==4.9.4
newspaper3k==0.2.8
trafilatura==1.6.4
readability-lxml==0.8.1

# Analysis
semantic-kernel==1.34.0
openai==1.67.0
numpy>=1.26.0

# System Monitoring
psutil>=6.1.1

# Validation
google-api-python-client==2.108.0

# Persistence & Reporting
pandas>=2.1.0
tabulate==0.9.0

# Azure Agents & Identity (for Bing Grounding)
azure-ai-projects
azure-ai-agents
azure-identity

# Chainlit Integration
chainlit>=1.0.0

# Fuzzy Matching for Company Resolution
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Additional Utilities
asyncio-throttle>=1.0.2

# Note: zoneinfo is built into Python 3.9+ for timezone handling
# If using Python < 3.9, install: pip install backports.zoneinfo 